task-#4 finished with library, clean up, testing

+++ StackExchangeClient/sqlite/src/main/java/ru/arturvasilov/stackexchangeclient/sqlite/table/BaseTable.java
@@ -25,7 +25,7 @@ public abstract class BaseTable<T> implements Table<T> {
 
     @Override
     public void onUpgrade(@NonNull SQLiteDatabase database, int oldVersion, int newVersion) {
-        if (newVersion <= getLastUpgradeVersion()) {
+        if (newVersion <= getLastUpgradeVersion() && newVersion > oldVersion) {
             database.execSQL("DROP TABLE IF EXISTS " + getTableName());
             onCreate(database);
         }
